refactor take card
