+++ src/index.css
@@ -3,7 +3,7 @@ html,body {
   height: 100%;
   margin: 0;
   padding: 0;
-  font-family: 'Roboto', sans-serif;
+  font-family: 'Noto Sans', sans-serif;
   box-sizing: border-box;
   background: #F6F9FE;
 }
