+++ src/js/pages/takes/TakeCard/Components/BottomButtons.js
