+++ src/js/pages/takes/TakeCard/Components/CommentRow.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
 import ReactPlayer from 'react-player';
-import PlayerTracker from '../../../../../components/PlayerTracker';
+import PlayerTracker from '../../../../components/PlayerTracker';
 import jdenticon from 'jdenticon';
-import config from '../../../../../../config/config';
+import config from '../../../../../config/config';
 
 export default class TakeCardCommentRow extends React.Component {
 
@@ -37,15 +37,15 @@ export default class TakeCardCommentRow extends React.Component {
     return (
       <CommentRow>
 
-        <IdenticonContainer>
-          <Identicon onClick={()=>this.play()} id={`CommentUser${comment.id}`} data-jdenticon-hash={comment.owner_icon_hash} />
-          <ReactPlayer url={`${config.streamingUrl}${comment.owner_name_audio}`} playing={this.state.playing} onEnded={()=> this.ended()}  />
-        </IdenticonContainer>
+
         <CommentPlayer >
           <PlayerTracker url={comment.location} />
         </CommentPlayer>
 
-        <RowButton> <i className="material-icons">delete</i> </RowButton>
+        <IdenticonContainer>
+          <Identicon onClick={()=>this.play()} id={`CommentUser${comment.id}`} data-jdenticon-hash={comment.owner_icon_hash} />
+          <ReactPlayer url={`${config.streamingUrl}${comment.owner_name_audio}`} playing={this.state.playing} onEnded={()=> this.ended()}  />
+        </IdenticonContainer>
 
       </CommentRow>
     );
@@ -57,10 +57,12 @@ export default class TakeCardCommentRow extends React.Component {
 const CommentRow = styled.div`
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
+  justify-content: space-evenly;
   align-items: center;
   border-bottom: solid 0.01vw lightgray;
   overflow: hidden;
+  background: white;
+  min-height: 40px;
 `;
 
 const CommentPlayer = styled.div`
@@ -69,29 +71,11 @@ const CommentPlayer = styled.div`
   flex: 1;
 `;
 
-
-const Button = styled.button`
-  font-size: 1.75vw;
-  flex: 1;
-  border: none;
-  align-self: stretch;
-  padding: 0.75vw;
-  border-top: solid 0.05vw #009CFF;
-  text-align:center;
-  background: none;
-`;
-
-const RowButton = styled(Button)`
-  flex:0;
-  padding: 0.4vw;
-  border-top: none;
-  font-size: 1vw;
-`;
-
 const IdenticonContainer =styled.div`
   background-color: white;
   height: 1.5vw;
   width: 1.5vw;
+  margin-right: 5%;
 
 `;
