+++ src/js/pages/takes/TakeCard/Components/TopBar.js
